+++ app/controllers/clearance/users_controller.rb
@@ -13,7 +13,7 @@ class Clearance::UsersController < ApplicationController
     @user = ::User.new(params[:user])
     if @user.save
       sign_in(@user)
-      flash_notice_after_create
+      flash_success_after_create
       redirect_to(url_after_create)
     else
       flash_failure_after_create
@@ -23,14 +23,14 @@ class Clearance::UsersController < ApplicationController
 
   private
 
-  def flash_notice_after_create
+  def flash_success_after_create
     flash[:notice] = translate(:signed_up,
       :scope   => [:clearance, :controllers, :users],
       :default => "You are now signed up.")
   end
 
   def flash_failure_after_create
-    flash.now[:failure] = translate(:bad_email_or_password,
+    flash.now[:notice] = translate(:bad_email_or_password,
       :scope   => [:clearance, :controllers, :passwords],
       :default => "Must be a valid email address. Password can't be blank.")
   end