+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * [#129] Denying access redirects to root_url when signed in, sign_in_url when signed out. (Dan Croak)
 * New configuration setting: denied_access_url. (Dan Croak)
+* Using flash :notice key everywhere now instead of :success and :failure. More in line with Rails conventions. (Dan Croak)
 
 0.11.2
 -------------------